Refactor scorers.py into scorers/ package
Split monolithic scorers.py into:
- scorers/core.py: all non-diffusion scorers and math helpers
- scorers/diffusion.py: DiffusionScorer and mu_distance
- scorers/__init__.py: score_rxns + re-exports for backward compat

Update import paths in recipe.py and get_scored_rxns.py to reference
the specific submodules.
