+++ src/rxn_ca/core/recipe.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass, field
 from .heating import HeatingSchedule
 from typing import Dict, List
 from ..utilities.helpers import format_chem_sys
-from ..reactions.scorers import TammanHuttigScoreErf, TammanHuttigScoreExponential, TammanHuttigScoreSoftplus, GibbsErfScore
+from ..reactions.scorers.core import TammanHuttigScoreErf, TammanHuttigScoreExponential, TammanHuttigScoreSoftplus, GibbsErfScore
 from ..phases.solid_phase_set import process_composition_dict, process_composition_list
 
 import json
