+++ src/rxn_ca/reactions/scorers/diffusion.py
@@ -1,144 +1,17 @@
 from __future__ import annotations
 
-import math
-from abc import ABC, abstractmethod
 from typing import Optional
 
 import numpy as np
 from pymatgen.core import Composition
-from tqdm import tqdm
-
 from rxn_network.entries.entry_set import GibbsEntrySet
 from rxn_network.reactions.computed import ComputedReaction
-from rxn_network.reactions.reaction_set import ReactionSet
 from rxn_network.thermo.chempot_diagram import ChemicalPotentialDiagram
 
-from ..phases.gasses import DEFAULT_GASES
-from ..phases.solid_phase_set import SolidPhaseSet
-from ..utilities.transport import TransportDatabase
-from .scored_reaction import ScoredReaction
-
-KB = 8.6173303e-5  # Boltzmann constant in eV/K
-Na = 6.02214076e23  # Avogadro's number
-
-
-def softplus(x):
-    return 1 / 3 * math.log(1 + math.exp(3 * x))
-
-
-def tamman_score_exp(t_tm_ratio):
-    return math.exp(4.82 * (t_tm_ratio) - 3.21)
-
-
-def tamman_score_softplus(t_tm_ratio):
-    return math.log(1 + math.exp(14 * (t_tm_ratio - 0.8)))
-
-
-def huttig_score_exp(t_tm_ratio):
-    return math.exp(2.41 * (t_tm_ratio) - 0.8)
-
-
-def huttig_score_softplus(t_tm_ratio):
-    return 0.25 * math.log(1 + math.exp(30 * (t_tm_ratio - 0.33)))
-
-
-def erf(x):
-    return 0.5 * (1 + math.erf(-35 * (x + 0.03)))
-
-
-class BasicScore(ABC):
-    def __init__(self, phase_set: SolidPhaseSet, temp: Optional[int] = None):
-        self.phases = phase_set
-        self.temp = temp
-
-    @abstractmethod
-    def score(self, rxn: ComputedReaction):
-        pass
-
-
-class TammanHuttigScoreExponential(BasicScore):
-    # https://en.wikipedia.org/wiki/Tammann_and_H%C3%BCttig_temperatures
-
-    def score(self, rxn: ComputedReaction):
-        phases = [c.reduced_formula for c in rxn.reactants]
-        non_gasses = [p for p in phases if p not in DEFAULT_GASES]
-        mps = [self.phases.get_melting_point(p) for p in non_gasses]
-        min_mp = min(mps)
-
-        delta_g_adjustment = softplus(-(2 * rxn.energy_per_atom + 0.8))
-
-        if len(non_gasses) < len(phases):
-            return huttig_score_exp(self.temp / min_mp) * delta_g_adjustment
-        return tamman_score_exp(self.temp / min_mp) * delta_g_adjustment
-
-
-class TammanHuttigScoreSoftplus(BasicScore):
-    # https://en.wikipedia.org/wiki/Tammann_and_H%C3%BCttig_temperatures
-
-    def score(self, rxn: ComputedReaction):
-        phases = [c.reduced_formula for c in rxn.reactants]
-        non_gasses = [p for p in phases if p not in DEFAULT_GASES]
-        mps = [self.phases.get_melting_point(p) for p in non_gasses]
-        min_mp = min(mps)
-
-        delta_g_adjustment = softplus(-(2 * rxn.energy_per_atom + 0.8))
-
-        if len(non_gasses) < len(phases):
-            return huttig_score_softplus(self.temp / min_mp) * delta_g_adjustment
-        return tamman_score_softplus(self.temp / min_mp) * delta_g_adjustment
-
-
-class TammanHuttigScoreErf(BasicScore):
-    # https://en.wikipedia.org/wiki/Tammann_and_H%C3%BCttig_temperatures
-
-    def score(self, rxn: ComputedReaction):
-        phases = [c.reduced_formula for c in rxn.reactants]
-        non_gasses = [p for p in phases if p not in self.phases.gas_phases]
-        mps = [self.phases.get_melting_point(p) for p in non_gasses]
-        min_mp = min(mps)
-
-        delta_g_adjustment = erf(rxn.energy_per_atom)
-
-        if len(non_gasses) == 1:
-            return huttig_score_softplus(self.temp / min_mp) * delta_g_adjustment
-        return tamman_score_softplus(self.temp / min_mp) * delta_g_adjustment
-
-
-class GibbsErfScore(BasicScore):
-    def score(self, rxn: ComputedReaction):
-        return erf(rxn.energy_per_atom)
-
-
-class TammanScore(BasicScore):
-    # https://en.wikipedia.org/wiki/Tammann_and_H%C3%BCttig_temperatures
-
-    def score(self, rxn: ComputedReaction):
-        phases = [c.reduced_formula for c in rxn.reactants]
-        non_gasses = [p for p in phases if p not in self.phases.gas_phases]
-        mps = [self.phases.get_melting_point(p) for p in non_gasses]
-        min_mp = min(mps)
-
-        delta_g_adjustment = erf(rxn.energy_per_atom)
-        return tamman_score_softplus(self.temp / min_mp) * delta_g_adjustment
-
-
-class ConstantScore(BasicScore):
-    def score(self, _):
-        return 1.0
-
-
-class TammanTightLinear(BasicScore):
-    def score(self, rxn: ComputedReaction):
-        phases = [c.reduced_formula for c in rxn.reactants]
-        non_gasses = [p for p in phases if p not in self.phases.gas_phases]
-        mps = [self.phases.get_melting_point(p) for p in non_gasses]
-        min_mp = min(mps)
-
-        def _score(x):
-            return 1 / 2 * (1 + math.erf(20 * (x - 0.6))) * (1 / 0.6 * x)
-
-        delta_g_adjustment = erf(rxn.energy_per_atom)
-        return _score(self.temp / min_mp) * delta_g_adjustment
+from ...phases.gasses import DEFAULT_GASES
+from ...phases.solid_phase_set import SolidPhaseSet
+from ...utilities.transport import TransportDatabase
+from .core import KB, Na, BasicScore, erf, softplus, tamman_score_softplus
 
 
 class DiffusionScorer(BasicScore):
@@ -188,7 +61,7 @@ class DiffusionScorer(BasicScore):
 
     def score(self, reaction: ComputedReaction):
         dG = reaction.energy_per_atom
-        
+
         phases = [c.reduced_formula for c in reaction.reactants]
         non_gasses = [p for p in phases if p not in DEFAULT_GASES]
         mps = [self.phase_set.get_melting_point(p) for p in non_gasses]
@@ -199,9 +72,9 @@ class DiffusionScorer(BasicScore):
         )
 
         onset_factor = tamman_score_softplus(self.temp / min_mp * self.tamman_factor)
-        
+
         if dG > 0.01:
-            return erf(dG)*onset_factor
+            return erf(dG) * onset_factor
 
         if len(reaction.reactants) > 1:
             reactants_dict = {c.reduced_formula: v for c, v in reaction.reactant_coeffs.items()}
@@ -308,22 +181,3 @@ def mu_distance(
         raise ValueError(f"Unknown mode={mode!r}")
 
     return {str(el): float(val) for el, val in zip(chempot.elements, mu_vec)}
-
-
-def score_rxns(
-    reactions: ReactionSet, scorer: BasicScore, phase_set: Optional[SolidPhaseSet] = None
-):
-    scored_reactions = []
-
-    for rxn in tqdm(
-        reactions.get_rxns(), desc=f"Scoring reactions... at temp {scorer.temp}"
-    ):
-        reactants = [r.reduced_formula for r in rxn.reactants]
-        non_gases = [r for r in reactants if r not in phase_set.gas_phases]
-        if len(non_gases) > 0:
-            scored_rxn = ScoredReaction.from_rxn_network(
-                scorer.score(rxn), rxn, phase_set.volumes
-            )
-            scored_reactions.append(scored_rxn)
-
-    return scored_reactions
